// src/app/shared/models/carga_academica.ts
export class CargaAcademica{  
  PROYECTO:  string;
  //ANIO:  string;
  //PER:  string;
  CODIGO_ASIGNATURA:  string;
  ASIGNATURA:  string;
  GRUPO:  string;
  //CODIGO_DOCENTE:  string;
  //DOCENTE:  string;
  //CODIGO_DIA:  string;
  DIA:  string;
  //CODIGO_HORA:  string;
  HORA:  string;
  //CODIGO_VINCULACION:  string;
  VINCULACION:  string;
  TIPO_VINCULACION:  string;
  //CODIGO_FACULTAD:  string;
  FACULTAD:  string;
  //CODIGO_DEDICACION:  string;
  DEDICACION:  string;
  //CODIGO_ACTIVIDAD:  string;
  //ACTIVIDAD:  string;
  NUMERO_ESTUDIANTES:  string;
  SALON:  string;
  
}